pyproject.toml
[tool.poetry]
name = "telegramvault"
version = "1.0.0"
description = "Telegram Intelligence Platform"
authors = ["TelegramVault"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
fastapi = "^0.112.0"
uvicorn = "0.23.2"
python-jose = "^3.5.0"
passlib = "^1.7.4"
bcrypt = "^5.0.0"
pydantic-settings = "^2.12.0"
python-multipart = "^0.0.21"
alembic = "^1.18.0"
sqlalchemy = "^2.0.45"
asyncpg = "^0.31.0"
aiofiles = "^25.1.0"
websockets = "^16.0"
pillow = "^12.1.0"
pytesseract = "^0.3.13"
opencv-python-headless = "^4.12.0.88"
telethon = "^1.42.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"